Fixed warnings and some file filters

+++ SP4/Base/Source/Player/Player.cpp
@@ -212,7 +212,7 @@ void CPlayer::Update(double dt, CTilemap* tile)
 		{
 			if (offSetDirectionY)
 			{
-				yOffSet += dt * 100;
+				yOffSet += static_cast<float>(dt * 100);
 				if (yOffSet > tile->GetTileSize())
 				{
 					yOffSet = 0;
@@ -222,7 +222,7 @@ void CPlayer::Update(double dt, CTilemap* tile)
 			}
 			else
 			{
-				yOffSet -= dt * 100;
+				yOffSet -= static_cast<float>(dt * 100);
 				if (yOffSet < -tile->GetTileSize())
 				{
 					yOffSet = 0;
@@ -235,7 +235,7 @@ void CPlayer::Update(double dt, CTilemap* tile)
 		{
 			if (offSetDirectionX)
 			{
-				xOffSet += dt * 100;
+				xOffSet += static_cast<float>(dt * 100);
 				if (xOffSet > tile->GetTileSize())
 				{
 					xOffSet = 0;
@@ -245,7 +245,7 @@ void CPlayer::Update(double dt, CTilemap* tile)
 			}
 			else
 			{
-				xOffSet -= dt * 100;
+				xOffSet -= static_cast<float>(dt * 100);
 				if (xOffSet < -tile->GetTileSize())
 				{
 					xOffSet = 0;
